Fix INSERT prepared statement support for SQLite
This also fixes CSV import in Kexi 3

SQLite is currently the only KDb driver that implements prepared statements entirely correctly by reusing the statement structures.
Here we are extending the prepared statement API so it's possible to specify that the SQL result structure is owned by the implementation and should not be deleted after each execution.
This is now the default behaviour. MySQL and PostgreSQL currently set the flag to false.

Also add some docs.