Expand scope of KDbTableSchemaChangeListener to allow registering queries for changes in table schema
Summary:
Listener related to a query can receive "close" request if table design the query depends on is about to be changed.
This way for example Kexi is able to ask users to close form or report if it uses query depending on altered table.

CCBUG:371896
FIXED-IN:3.1.0

When table schema is removed unregister it from KDbTableSchemaChangeListener objects

FIXED-IN:3.1.0

Test Plan: Test Kexi for https://phabricator.kde.org/D8118

Reviewers: piggz

Reviewed By: piggz

Tags: #kdb

Maniphest Tasks: T7703, T7704

Differential Revision: https://phabricator.kde.org/D8117