Expand feature test for buffered cursors
Summary:
When testing a driver’s cached cursors it is interesting to see whether
the cursor actually can go back to the first record and then return the
values for all of them again.

Checking only the return value of KDbCursor::moveFirst and
KDbCursor::moveLast is not enough to see if it works, because they just
return true even though the driver does not change anything.

Reviewers: staniek, piggz

Subscribers: Kexi-Devel-list

Tags: #kdb

FIXED-IN:3.2.1

Differential Revision: https://phabricator.kde.org/D29185